fix servo attach position not b16 and arduino createAndStart
not fully reentrant

+++ src/resource/Arduino/MRLComm/MrlComm.h
@@ -114,8 +114,8 @@ public:
 	void setTrigger( byte pin,  byte triggerValue);
 	// > setDebounce/pin/delay
 	void setDebounce( byte pin,  byte delay);
-	// > servoAttach/deviceId/pin/initPos/b16 initVelocity
-	void servoAttach( byte deviceId,  byte pin,  byte initPos,  int initVelocity);
+	// > servoAttach/deviceId/pin/b16 initPos/b16 initVelocity
+	void servoAttach( byte deviceId,  byte pin,  int initPos,  int initVelocity);
 	// > servoAttachPin/deviceId/pin
 	void servoAttachPin( byte deviceId,  byte pin);
 	// > servoDetachPin/deviceId
